refactor: move tests to the proper modules
