+++ src/num_of_pages.rs
@@ -122,3 +122,138 @@ impl<T: PageSize> DivAssign<usize> for NumOfPages<T> {
         *self = *self / rhs;
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::NumOfPages;
+    use x86_64::structures::paging::Size1GiB;
+    use x86_64::structures::paging::Size2MiB;
+    use x86_64::structures::paging::Size4KiB;
+
+    #[test]
+    fn get_value_from_num_of_pages() {
+        let pages = NumOfPages::<Size4KiB>::new(334);
+        assert_eq!(pages.as_usize(), 334);
+    }
+
+    #[test]
+    fn pages_to_bytes_4k() {
+        let num_of_pages = NumOfPages::<Size4KiB>::new(1);
+        assert_eq!(num_of_pages.as_bytes().as_usize(), 0x1000);
+    }
+
+    #[test]
+    fn pages_to_bytes_2m() {
+        let num_of_pages = NumOfPages::<Size2MiB>::new(1);
+        assert_eq!(num_of_pages.as_bytes().as_usize(), 0x200000);
+    }
+
+    #[test]
+    fn pages_to_bytes_1g() {
+        let num_of_pages = NumOfPages::<Size1GiB>::new(1);
+        assert_eq!(num_of_pages.as_bytes().as_usize(), 0x40000000);
+    }
+
+    #[test]
+    fn addition_pages_to_pages() {
+        let p1 = NumOfPages::<Size4KiB>::new(3);
+        let p2 = NumOfPages::<Size4KiB>::new(1);
+        let sum = p1 + p2;
+
+        assert_eq!(sum.as_usize(), 4);
+    }
+
+    #[test]
+    fn add_usize_to_num_of_pages() {
+        let n = NumOfPages::<Size4KiB>::new(3);
+
+        assert_eq!(n + 7, NumOfPages::new(10));
+    }
+
+    #[test]
+    fn subtraction_pages_from_pages() {
+        let p1 = NumOfPages::<Size4KiB>::new(3);
+        let p2 = NumOfPages::<Size4KiB>::new(1);
+        let diff = p1 - p2;
+
+        assert_eq!(diff.as_usize(), 2);
+    }
+
+    #[test]
+    fn subtract_usize_from_num_of_pages() {
+        let n = NumOfPages::<Size4KiB>::new(5);
+
+        assert_eq!(n - 3, NumOfPages::new(2));
+    }
+
+    #[test]
+    fn add_assign_pages_to_pages() {
+        let mut p1 = NumOfPages::<Size4KiB>::new(3);
+        p1 += NumOfPages::<Size4KiB>::new(1);
+
+        assert_eq!(p1.as_usize(), 4);
+    }
+
+    #[test]
+    fn add_assign_usize_to_pages() {
+        let mut p1 = NumOfPages::<Size4KiB>::new(3);
+        p1 += 1;
+
+        assert_eq!(p1.as_usize(), 4);
+    }
+
+    #[test]
+    fn sub_assign_pages_to_pages() {
+        let mut p1 = NumOfPages::<Size4KiB>::new(3);
+        p1 -= NumOfPages::<Size4KiB>::new(1);
+
+        assert_eq!(p1.as_usize(), 2);
+    }
+
+    #[test]
+    fn sub_assign_usize_to_num_of_pages() {
+        let mut p1 = NumOfPages::<Size4KiB>::new(10);
+        p1 -= 3;
+
+        assert_eq!(p1, NumOfPages::new(7));
+    }
+
+    #[test]
+    fn mul_pages_by_usize() {
+        let p = NumOfPages::<Size4KiB>::new(3);
+        let mul = p * 4;
+
+        assert_eq!(mul.as_usize(), 12);
+    }
+
+    #[test]
+    fn mul_assign_pages_by_usize() {
+        let mut p = NumOfPages::<Size4KiB>::new(3);
+        p *= 4;
+
+        assert_eq!(p.as_usize(), 12);
+    }
+
+    #[test]
+    fn div_num_of_pages_by_usize() {
+        let p1 = NumOfPages::<Size4KiB>::new(3);
+        let div = p1 / 2;
+
+        assert_eq!(div.as_usize(), 1);
+    }
+
+    #[test]
+    fn divassign_num_of_pages_by_usize() {
+        let mut p = NumOfPages::<Size4KiB>::new(3);
+        p /= 2;
+
+        assert_eq!(p.as_usize(), 1);
+    }
+
+    #[test]
+    fn num_of_pages_zero() {
+        let n = NumOfPages::<Size4KiB>::zero();
+
+        assert_eq!(n.as_usize(), 0);
+    }
+}
